# packages/contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/disputes": {
      "get": {
        "operationId": "AdminDisputesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List disputes",
        "tags": [
          "Disputes"
        ]
      },
      "post": {
        "operationId": "AdminDisputesController_open",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenDisputeBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Open a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/admin/disputes/{id}": {
      "get": {
        "operationId": "AdminDisputesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get dispute detail",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/admin/disputes/{id}/assign": {
      "patch": {
        "operationId": "AdminDisputesController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignDisputeBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/admin/disputes/{id}/status": {
      "patch": {
        "operationId": "AdminDisputesController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeStatusChangeBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change dispute status",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/admin/disputes/{id}/messages": {
      "post": {
        "operationId": "AdminDisputesController_addMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisputeMessageBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a message to a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
    "/api/v1/admin/disputes/{id}/files": {
      "post": {
        "operationId": "AdminDisputesController_attachFile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachDisputeFileBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Attach an evidence file to a dispute",
        "tags": [
          "Disputes"
        ]
      }
    },
...
    "/api/v1/disputes/{id}": {
      "get": {
        "operationId": "DisputesController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Follow one of my disputes",
        "tags": [
          "Disputes"
        ]
      }
    },
...
    "/api/v1/admin/notifications": {
      "get": {
        "operationId": "AdminNotificationsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/admin/notifications/templates": {
      "get": {
        "operationId": "AdminNotificationsController_templates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List notification templates",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/admin/notifications/send": {
      "post": {
        "operationId": "AdminNotificationsController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationBodyDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a system notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "PRESTGO Back-office Admin and Backend Services",
    "description": "Versioned contract for PRESTGO V1 backend and internal admin workflows.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OpenDisputeBodyDto": {
        "type": "object",
        "properties": {}
      },
      "AssignDisputeBodyDto": {
        "type": "object",
        "properties": {}
      },
      "DisputeStatusChangeBodyDto": {
        "type": "object",
        "properties": {}
      },
      "DisputeMessageBodyDto": {
        "type": "object",
        "properties": {}
      },
      "AttachDisputeFileBodyDto": {
        "type": "object",
        "properties": {}
      },
...
      "SendNotificationBodyDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}